backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.10
python-multipart
requests
httpx